Feature/bc 704 hifi comments (#120)
[BC-704: hifi
comments](https://bootcamper.atlassian.net/browse/BC-704?atlOrigin=eyJpIjoiOTM4YzcwZmMxZDUyNDIwODk1MmRmZmJkMmRmY2Q3NDYiLCJwIjoiaiJ9)

Includes: 
* deletes a bunch of deprecated/unused files & relocates some to the
proper folders
* creates separate files for comments
* removes dead comments

---------
