--- controllers/chats.js
@@ -1,3 +1,0 @@
-// Potential Chat Controllers
-  // Create New Chat
-  // Get All Chats for a Project
