+++ controllers/project/projects.js
@@ -1,17 +1,8 @@
-/**
- * todo:
- * - need to add edge cases
- * - need to add edge cases
- * - refactor this code because there is a for sure cleaner way to do all of this
- *  - might even be able to make a helper function since a lot of the special endpoints share the same functionality
- * - consider what properties containing objectId's should be populated in the response
- */
 import Project from '../../models/project.js';
 import User from '../../models/user.js';
 import { findCommonAvailability } from '../../utils/availability.js';
 import { convertQueryAttributesToMongoString } from '../../utils/helperFunctions.js';
 
-//basic CRUD functions:
 export const getAllProjects = async (req, res) => {
   try {
     const projects = await Project.find();
@@ -44,7 +35,7 @@ export const getOneProject = async (req, res) => {
     console.log(error.message);
     res.status(500).json({ error: error.message });
   }
-}; // tested and is good
+};
 
 export const getProjectMembers = async (req, res) => {
   try {
@@ -69,7 +60,7 @@ export const createProject = async (req, res) => {
     console.log(error.message);
     res.status(500).json({ error: error.message });
   }
-}; // works, requires userID
+};
 
 export const updateProject = async (req, res) => {
   try {
@@ -90,7 +81,7 @@ export const updateProject = async (req, res) => {
     console.log(error.message);
     res.status(500).json({ error: error.message });
   }
-}; // works
+};
 
 export const deleteProject = async (req, res) => {
   try {
@@ -104,7 +95,7 @@ export const deleteProject = async (req, res) => {
     console.log(error.message);
     res.status(500).json({ error: error.message });
   }
-}; // works
+};
 
 export const updateUserAndProject = async (req, res) => {
   try {
@@ -112,7 +103,7 @@ export const updateUserAndProject = async (req, res) => {
     const { userUpdate, userID } = req.body.user;
     const project = await Project.findByIdAndUpdate(projectID, projectUpdate, {
       new: true,
-    }); // .populate({ path: "team_members", model: User }); //review
+    });
     const user = await User.findByIdAndUpdate(userID, userUpdate, {
       new: true,
     });
@@ -127,15 +118,12 @@ export const getTeamCommonAvailability = async (req, res) => {
   try {
     const { projectId } = req.params;
     const project = await Project.findById(projectId)
-      .populate([
-        { path: 'members.engineers' },
-        { path: 'members.designers' },
-      ])
+      .populate([{ path: 'members.engineers' }, { path: 'members.designers' }])
       .exec();
 
     if (project) {
-      const members = [...project.members.engineers, ...project.members.designers]
-      const commonAvailability = findCommonAvailability(members)
+      const members = [...project.members.engineers, ...project.members.designers];
+      const commonAvailability = findCommonAvailability(members);
 
       return res.json(commonAvailability);
     }
@@ -145,11 +133,3 @@ export const getTeamCommonAvailability = async (req, res) => {
     res.status(500).json({ error: error.message });
   }
 };
-
-// Potential Project new controllers
-// Add members to project (similar to updateUserAndProject?)
-// Add Task to Project
-// Delete Task from Project (or better to 'archive' and just update the task status?)
-// Add Meeting Id to Project
-// Delete Meeting from Project (better to delete or archive?)
-// Add Chat Id to Project
