--- controllers/project/tasks.js
@@ -1,5 +1,0 @@
-// Potential Task Controllers
-//
-// Create New Task
-// Update Task Status (includes archive aka delete)
-// Update Task Assignee
