+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bootcamper",
-  "version": "1.0.5",
+  "version": "1.0.6",
   "description": "",
   "main": "index.js",
   "scripts": {
