--- routes/chats.js
@@ -1,10 +1,0 @@
-import { Router } from 'express';
-// import controllers
-
-const router = Router();
-
-// chat routes
-// CREATE new chat
-// GET All Chats for a Project
-
-export default router
