+++ routes/index.js
@@ -3,6 +3,7 @@ import { Router } from 'express';
 import authRoutes from './user/auth.js';
 import availabilityRoutes from './calendar/availability.js';
 import calendarRoutes from './calendar/calendars.js';
+import commentRoutes from './project/comments.js';
 import chatThreadRoutes from './chat/thread.js';
 import eventRoutes from './calendar/events.js';
 import groupChatRoutes from './chat/groupChat.js';
@@ -24,6 +25,7 @@ router.get('/', (req, res) => res.send('api root'));
 router.use('/', authRoutes);
 router.use('/', availabilityRoutes);
 router.use('/', calendarRoutes);
+router.use('/', commentRoutes);
 router.use('/', chatThreadRoutes);
 router.use('/', eventRoutes);
 router.use('/', groupChatRoutes);
